refactor: move wallet helper use-cases into 'private' sub-dir
