+++ src/app/wallets/get-pending-onchain-balance-for-wallet.ts
@@ -5,7 +5,7 @@ import { TxFilter } from "@domain/bitcoin/onchain"
 import { baseLogger } from "@services/logger"
 import { IncomingOnChainTxHandler } from "@domain/bitcoin/onchain/incoming-tx-handler"
 
-import getOnChainTxs from "./get-on-chain-txs"
+import { getOnChainTxs } from "./private"
 
 export const getPendingOnChainBalanceForWallets = async (
   wallets: Wallet[],
