+++ src/app/wallets/index.ts
@@ -1,18 +1,18 @@
+export * from "./add-invoice-for-wallet"
 export * from "./create-on-chain-address"
 export * from "./get-balance-for-wallet"
-export * from "./get-pending-onchain-balance-for-wallet"
-export * from "./get-transactions-for-wallet"
 export * from "./get-last-on-chain-address"
 export * from "./get-on-chain-fee"
+export * from "./get-pending-onchain-balance-for-wallet"
 export * from "./get-transaction-by-id"
-export * from "./get-transactions-by-hash"
 export * from "./get-transactions-by-addresses"
+export * from "./get-transactions-by-hash"
+export * from "./get-transactions-for-wallet"
+export * from "./reimburse-failed-usd"
+export * from "./reimburse-fee"
+export * from "./send-on-chain"
 export * from "./update-on-chain-receipt"
 export * from "./update-pending-invoices"
-export * from "./add-invoice-for-wallet"
-export * from "./reimburse-fee"
-export * from "./reimburse-failed-usd"
-export * from "./send-on-chain"
 
 import { WalletsRepository } from "@services/mongoose"
